divided header components
