+++ frontend/src/types/menu.ts
@@ -1,0 +1,4 @@
+export interface MenuItem  {
+    name: string;
+    link: string;
+};
